Collision Operator

Single Band and Naïve Multiband

For a single band with a momentum independent scattering potential $V(\mathbf{q}) = V$,

\[ \mathbf{L}_{ij} = \frac{1}{\d A_i} \frac{2\pi}{1 - f^{(0)}_i} |V|^2 \frac{1}{(2\pi)^6} \sum_{m} (f^{(0)}_j (1 - f^{(0)}_m) \mathcal{K}_{ijm} - 2f^{(0)}_m (1 - f^{(0)}_j)\mathcal{K}_{imj}) \]

where

\[ \mathcal{K}_{ijm} = \int_i d^2 \mathbf{k}_i \int_j d^2 \mathbf{k}_j \int_m d^2 \mathbf{k}_m (1 - f^{(0)}(\mathbf{k}_i + \mathbf{k}_j - \mathbf{k}_m)) \delta(\varepsilon_i + \varepsilon_j - \varepsilon_m - \varepsilon_\mu(\mathbf{k}_i + \mathbf{k}_j - \mathbf{k}_m)).\]

A first attempt at incorporating interband scattering was to perform a sum over the bands for the energy corresponding the final momentum computed by enforcing momentum conservation:

Ludwig.Γabc! — Function

Γabc!(ζ, a, b, c, T, Δε, hams::Vector{Function})

Compute the integral

\[ \mathcal{K}^\mu_{abc} \equiv \sum_{\mu} \int_a \int_b \int_c (1 - f^{(0)}(\mathbf{k}_a + \mathbf{k}_b + \mathbf{k}_c)) \delta(\varepsilon_a + \varepsilon_b - \varepsilon_c - \varepsilon^\mu(\mathbf{k}_a + \mathbf{k}_b - \mathbf{k}_c)) \]

where $\mu$ denotes the band index of the dispersion in hams and

\[ \int_i \equiv \frac{1}{a^2} \int_{\mathbf{k} \in \mathcal{P}_i} d^2\mathbf{k}\]

is an integral over momenta in patch $\mathcal{P}_i$.

Γabc!(ζ, a, b, c, T, Δε, h::Function)

When only a single band h is provided, compute the integration kernel without summing over bands.

Γabc!(ζ, a, b, c, T, Δε, itps::Vector{ScaledInterpolation})

The same integral where energies are evaluated by interpolation of the bands over the 1st Brillouin Zone

The collision matrix can then be populated in this simple model using

Ludwig.electron_electron! — Method

electron_electron!(L, grid, Δε, T, hams::Vector{Function})

Populate the collision matrix L defined on patches in grid assuming equal weight for interband scattering given explicit functions for each band in hams.

or

Ludwig.electron_electron! — Method

electron_electron!(L, grid, Δε, T, itps::Vector{ScaledInterpolation})

Populate the collision matrix L defined on patches in grid assuming equal weight for interband scattering given gridded interpolations of bands in itps.

Improved Multiband

The above model is not realistic, however. The bands consist of hybridized orbitals which have different overlap. To account for this, the next simplest model one can propose for the scattering term is

\[ U\sum_{i} \sum_{a,b} n_{i,a} n_{i,b}\]

where $i$ represents a site index and $a, b$ are orbital indices. To handle this perturbation in our framework of scattering using the Born approximation, we need to evaluate this interaction term in the eigenbasis of the bare Hamiltonian.

\[\begin{aligned} \sum_i n_{i,a} n_{i, b} &= \sum_{i}c^\dagger_{i,a} c_{i,a} c^\dagger_{i,b} c_{i,b}\\ &= \frac{1}{N^2} \sum_{\mathbf{k}_1, \mathbf{k}_2, \mathbf{k}_3, \mathbf{k}_4} \left( e^{i (\mathbf{k}_3 + \mathbf{k}_4 - \mathbf{k}_1 - \mathbf{k}_2) \cdot \mathbf{R}_i}\right ) c^\dagger_{\mathbf{k}_3,a} c_{\mathbf{k}_1,a} c^\dagger_{\mathbf{k}_4,b} c_{\mathbf{k}_2,b}\\ &= \frac{1}{N^2} \sum_{\mathbf{k}_1, \mathbf{k}_2, \mathbf{k}_3, \mathbf{k}_4} N \delta_{\mathbf{k}_4,\mathbf{k}_1 + \mathbf{k}_2 - \mathbf{k}_3} c^\dagger_{\mathbf{k}_3,a} c_{\mathbf{k}_1,a} c^\dagger_{\mathbf{k}_4,b} c_{\mathbf{k}_2,b}\\ &= \frac{1}{N} \sum_{\mathbf{k}_1, \mathbf{k}_2, \mathbf{q}} c^\dagger_{\mathbf{k}_1 - \mathbf{q},a} c_{\mathbf{k}_1,a} c^\dagger_{\mathbf{k}_2 + \mathbf{q},b} c_{\mathbf{k}_2,b}\\ &= \frac{1}{N} \sum_{\mathbf{k}_1, \mathbf{k}_2, \mathbf{q}} \sum_{\mu\nu\sigma\tau} c^\dagger_{\mathbf{k}_1 - \mathbf{q},\sigma} \left( W_{\mathbf{k}_1 - \mathbf{q}}^{a \sigma}\right)^* W_{\mathbf{k}_1}^{a \mu} c_{\mathbf{k}_1, \mu} c^\dagger_{\mathbf{k}_2 + \mathbf{q},\tau} \left( W_{\mathbf{k}_2 + \mathbf{q}}^{b \tau}\right)^* W_{\mathbf{k}_2}^{b \nu} c_{\mathbf{k}_2, \nu} \end{aligned}\]

From the above, we can define the interband connection as

\[ F_{\mathbf{k}_1, \mathbf{k}_2}^{\mu\nu} \equiv \sum_a \left(W_{\mathbf{k}_1}^{a \mu} \right)^* W_{\mathbf{k}_2}^{a \nu} = \sum_a \left(W^\dagger_{\mathbf{k}_1}\right)^{\mu a} W_{\mathbf{k}_2}^{a \nu} = \left(W^\dagger_{\mathbf{k}_1} W_{\mathbf{k}_2} \right)^{\mu \nu } \]

so the interaction term becomes

\[U\sum_{i} \sum_{a,b} n_{i,a} n_{i,b} = \frac{U}{N} \sum_{\mathbf{k}_1, \mathbf{k}_2, \mathbf{q}} \sum_{\mu\nu\sigma\tau} c^\dagger_{\mathbf{k}_1 - \mathbf{q},\sigma} F^{\sigma\mu}_{\mathbf{k}_1 - \mathbf{q}, \mathbf{k}_1} c_{\mathbf{k}_1, \mu} c^\dagger_{\mathbf{k}_2 + \mathbf{q},\tau} F^{\tau\nu}_{\mathbf{k}_2 + \mathbf{q}, \mathbf{k}_2} c_{\mathbf{k}_2, \nu}.\]

The collision matrix elements can then be populated using

Ludwig.electron_electron! — Method

electron_electron!(L, grid, Δε, T, H, N, cf_eigvecs!)

Populate the collision matrix L defined on patches in grid for a scattering perturbation to the Hamiltonian of the form

V = \frac{U}{N} \sum_{a,b} \sum_{\mathbf{k}_1, \mathbf{k}_2, \mathbf{q}} \sum_{\mu\nu\sigma\tau} c^\dagger_{\mathbf{k}_1 - \mathbf{q},\sigma} \left( W_{\mathbf{k}_1 - \mathbf{q}}^{a \sigma}\right)^* W_{\mathbf{k}_1}^{a \mu} c_{\mathbf{k}_1, \mu} c^\dagger_{\mathbf{k}_2 + \mathbf{q},\tau} \left( W_{\mathbf{k}_2 + \mathbf{q}}^{b \tau}\right)^* W_{\mathbf{k}_2}^{b \nu} c_{\mathbf{k}_2, \nu}

The field w of the type Patch gives the column of $W_\mathbf{k}$ corresponding to the field band_index evaluated at the field band_index.

The function cf_eigvecs! is a convenience function that speeds up performance by populating a pre-allocated matrix with the closed form solution for the eigenvectors of H.

which is threaded to improve performance and uses the alternate kernel function ee_kernel! instead of Γabc! for improved memory allocation handling when dealing with the larger memory requirements of this interband model.

Ludwig.ee_kernel! — Function

ee_kernel!(K, grid, i, T, Δε, itps, f0s)

Populate the preallocated array K with the values of $\mathcal{K}^\mu_{ijm}$ for the row i of the collision operator.

Energies are calculated from itps, the interpolation of each band and f0s is a vector of the Fermi-Dirac distribution evaluated for the energy of each patch in grid at temperature T.