Fermi Surface Centered Meshes

Ludwig.Patch — Type

Representation of a patch in momentum space to be integrated over when calculating the collision integral kernel.

Fields

momentum: Momentum in 1st BZ scaled by $2\pi / a$

energies: Eigenvalues of Hamiltonian evaluated at momentum

band_index: Index of the Fermi surface from which the patch was generated

v: The group velocity at momentum taking energies[band_index] as the dispersion

dV: Area of the patch in units of $(2\pi/a)^2$

jinv: Jacobian of transformation from $(k_x, k_y) \mapsto (\varepsilon, \theta)$, the local patch coordinates

w: The weights of the original orbital basis corresponding to the band_indexth eigenvalue

corners: Indices of coordinates in parent Mesh struct of corners for plotting