Fermi Surface Centered Meshes

At finite temperature, the particles participating in scattering with non-negligible weight belong to a narrow annulus of energies near the Fermi surface. Thus, for calculating the Boltzmann collision matrix, the sampled momenta are chosen to be uniformly distributed in arclength along energy contours and energy above and below the Fermi surface. This follows the approach outlined in J. M. Buhmann's PhD Thesis.

The sampled momenta lie at the local energy-arclength coordinate center of each Patch, and information about the patch used in integration is stored in the fields of a variable of type Patch.

Ludwig.FSMesh.Patch — Type

Patch(e::Float64, k::SVector{2,Float64}, v::SVector{2,Float64}, de::Float64, dV::Float64, jinv::Matrix{Float64}, djinv::Float64, band_index::Int)

Construct a `Patch' object defining regions of momentum space over which to integrate.

Fields

e: energy

k: momentum

v: group velocity

de: width of patch in energy

dV: area of patch in momentum space

jinv: Jacobian of transformation from (kx, ky) –> (E, s)

djinv: determinant of above Jacobian

band_index: index of band from which e was sampled at k

Patches within a mesh are generated by creating a contour at a sample energy $\varepsilon$ and two additional contours at $\pm\Delta\varepsilon$. The central contour is cut into segments of equal arclength, and the intersection of the gradient of energy at each cut with the boundary contours at $\pm\Delta\varepsilon$ define the corners of the patch. Taking the representative momentum to lie halfway along the central contour between two cuts means we have a patch which when mapped to local energy-arclength coordinates is rectangular with area $\Delta \varepsilon \Delta s$ and the representative momentum of the patch at the origin. This scheme is ideal for integration.

These patches are stored in a container type Mesh which contains additional fields for plotting.

Ludwig.FSMesh.Mesh — Type

Mesh(patches::Vector{Patch}, corners::Vector{SVector{2,Float64}}, corner_inds::Vector{SVector{4,Int}}})

Construct a container struct of patches which contains information for plotting functions defined on patch centers.

Fields

patches: Vector of patches

corners: Vector of points on patch corners for plotting mesh

corner_inds: Vector of vector of indices in the corners vector corresponding to the corners of each patch in patches

References

J. M. Buhmann, Unconventional Transport Properties of Correlated Two-Dimensional Fermi Liquids, Ph.D. thesis, Institute for Theoretical Physics ETH Zurich (2013).