Collision Operator

Ludwig.Γabc! — Function

Γabc!(ζ, a, b, c, T, Δε, hams::Vector{Function})

Compute the integral

\[ \Gamma_{abc} \equiv \sum_{\mu} \int_a \int_b \int_c (1 - f^{(0)}(\mathbf{k}_a + \mathbf{k}_b + \mathbf{k}_c)) \delta(\varepsilon_a + \varepsilon_b - \varepsilon_c - \varepsilon^\mu(\mathbf{k}_a + \mathbf{k}_b - \mathbf{k}_c)) \]

where $\mu$ denotes the band index of the dispersion in hams and

\[ \int_i \equiv \frac{1}{a^2} \int_{\mathbf{k} \in \mathcal{P}_i} d^2\mathbf{k}\]

is an integral over momenta in patch $\mathcal{P}_i$.

Γabc!(ζ, a, b, c, T, Δε, h::Function)

When only a single band h is provided, compute the integration kernel without summing over bands.

Γabc!(ζ, a, b, c, T, Δε, itps::Vector{ScaledInterpolation})

The same integral where energies are evaluated by interpolation of the bands over the 1st Brillouin Zone