Fermi Surface Centered Meshes

At finite temperature, the particles participating in scattering with non-negligible weight belong to a narrow annulus of energies near the Fermi surface. Thus, for calculating the Boltzmann collision matrix, the sampled momenta are chosen to be uniformly distributed in angle and energy between a temperature-dependent threshhold. This follows the approach outlined in J. M. Buhmann's PhD Thesis.

The sampled momenta lie at the orthocenter of each Patch, and information about the patch used in integration is stored in the fields of a variable of type Patch.

Ludwig.Patch — Type

Representation of a patch in momentum space to be integrated over when calculating the collision integral kernel.

Fields

momentum: Momentum in 1st BZ scaled by $2\pi / a$

energies: Eigenvalues of Hamiltonian evaluated at momentum

band_index: Index of the Fermi surface from which the patch was generated

v: The group velocity at momentum taking energies[band_index] as the dispersion

dV: Area of the patch in units of $(2\pi/a)^2$

jinv: Jacobian of transformation from $(k_x, k_y) \mapsto (\varepsilon, \theta)$, the local patch coordinates

w: The weights of the original orbital basis corresponding to the band_indexth eigenvalue

corners: Indices of coordinates in parent Mesh struct of corners for plotting

Each patch belongs to a specific band, and the following method allows the energy of that patch to be accessed without storing this as a separate field.

Ludwig.energy — Method

energy(p::Patch)

Return the energy corresponding to the band for which p was generated.

These patches are stored in a container type Mesh which contains additional fields for plotting.

Ludwig.Mesh — Type

Container struct for patches over which to integrate.

Fields

patches: Vector of patches

corners: Vector of points on patch corners for plotting mesh

n_bands: Dimension of the generating Hamiltonian

References

J. M. Buhmann, Unconventional Transport Properties of Correlated Two-Dimensional Fermi Liquids, Ph.D. thesis, Institute for Theoretical Physics ETH Zurich (2013).